Further small renaming / refactoring changes
