+++ src/gt4py/cartesian/__init__.py
@@ -17,11 +17,6 @@
 import typing
 
 
-__copyright__: typing.Final = "Copyright (c) 2014-2022 ETH Zurich"
-__license__: typing.Final = "GPLv3+"
-
-from .version import __version__, __version_info__  # isort:skip
-
 from . import (
     caching,
     cli,
